requirements.txt
-f /usr/share/pip-wheels
backcall==0.1.0
boto3==1.9.91
botocore==1.12.91
cloudpickle==0.7.0
cycler==0.10.0
dask==1.1.1
decorator==4.3.2
Django==2.1.5
django-crispy-forms==1.7.2
django-storages==1.7.1
docutils==0.14
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
jmespath==0.9.3
kiwisolver==1.0.1
matplotlib==3.0.2
mysql==0.0.2
mysqlclient==1.4.1
networkx==2.2
numpy==1.16.1
opencv-python==4.0.0.21
parso==0.3.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.8
ptyprocess==0.6.0
pygame==1.9.4
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
PyWavelets==1.0.1
s3transfer==0.2.0
scikit-image==0.14.2
scipy==1.2.1
six==1.12.0
toolz==0.9.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
